Posibilidad de crear notas

+++ app/src/main/java/com/bios/android/notesapp/dto/Note.java
@@ -3,8 +3,8 @@ package com.bios.android.notesapp.dto;
 public class Note {
     private String name;
     private String content;
-    private Long id_user;
-    private Long id;
+    private Integer id_user;
+    private Integer id;
 
     public String getName() {
         return name;
@@ -22,19 +22,19 @@ public class Note {
         this.content = content;
     }
 
-    public Long getId_user() {
+    public Integer getId_user() {
         return id_user;
     }
 
-    public void setId_user(Long id_user) {
+    public void setId_user(Integer id_user) {
         this.id_user = id_user;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 }
